Move the piece file to the pieces dir
