+++ src/pieces/types/mod.rs
@@ -2,7 +2,6 @@ pub const BOARD_SIZE: usize = 8;
 
 pub mod color;
 pub mod move_direction;
-pub mod piece;
 pub mod position;
 
 #[cfg(test)]
